chore(scripts): move esbuild config into scripts directory
