+++ biome.jsonc
@@ -117,7 +117,6 @@
   "overrides": [
     {
       "includes": [
-        "esbuild/**/*",
         "scripts/**/*",
         ".github/actions/**/*",
         "src/content/debug.ts"
